Move vscode dependent file out to get tests passing
